Separate datastructs into their own testing packages
