+++ src/test/java/com/kgarsjo/datastructs/stack/EasyStackTest.java
@@ -1,6 +1,5 @@
-package com.kgarsjo.datastructs.tests;
+package com.kgarsjo.datastructs.stack;
 
-import com.kgarsjo.datastructs.stack.EasyStack;
 import org.junit.Before;
 
 public class EasyStackTest extends AbstractStackTest {
