+++ src/test/java/com/kgarsjo/datastructs/stack/LinkedStackTest.java
@@ -1,6 +1,5 @@
-package com.kgarsjo.datastructs.tests;
+package com.kgarsjo.datastructs.stack;
 
-import com.kgarsjo.datastructs.stack.LinkedStack;
 import org.junit.Before;
 
 public class LinkedStackTest extends AbstractStackTest {
